==> 1_Cleaning dataset.ipynb <==
import pandas as pd
import urllib
import os
import string
import re
import numpy as np

df = pd.read_csv("Data/voies_paris_labelisé.csv", sep=';', index_col=0)
df.head()

len(df)

# 1. Clean "HISTORIQUE"

On va utiliser la variable "HISTORIQUE", pour extraire des informations relatives à la personne qui a donné son nom à la voie.

def cleaning_historique() :
    df['Libellé VOIE'] = df['Libellé VOIE'].astype(str).apply(lambda x: x.replace('-', ' '))
    df['Libellé VOIE'] = df['Libellé VOIE'].astype(str).apply(lambda x: x.replace('et ', ''))
    
    # Removing expressions "Origin. du nom"
    removed_expressions = ["Orig. du nom. ~ ", "Orig.du nom.~ ", "orig. du nom. ~", "orig. du nom.~ ", "Orig. Du nom.~", "Orig. Du nom. ~"]
    for expression in removed_expressions :
        df["HISTORIQUE"] = df['HISTORIQUE'].astype(str).apply(lambda x: x.replace(expression, ''))

cleaning_historique()

# 2. Voisinages

On distingue les rues qui portent des noms d'individus et les autres.

Plusieurs façons de reconnaître les rues qui portent des noms d'individus :
- On a souvent les dates de naissance et mort entre parenthèses pour les rues qui portent des noms de personnes
- S'il s'agit d'un propriétaire (qui possédait donc des immeubles dans cette rue), on peut filter avec "propriétaire"
- Pour certaines rues (plutôt des impasses), on n'a pas la description de la personne, car il s'agit d'une rue au voisinage d'un boulevard par exemple (et l'information sur la personne est disponible à la ligne correspondant au boulevard)
- Ou sinon, si on n'a ni l'un ni l'autre, on peut utiliser la liste des prénoms (puisque pour certains individus, on n'a ni les autres informations ci-dessus)

# dataframe voisinages: on remplace la "Rues" par la vraie référence à la rue
voisinages = df.loc[(df['HISTORIQUE'].str.startswith("Voisinage")) | (df['HISTORIQUE'].str.startswith("Voir")) | (df['Libellé VOIE'] == 'boulevard Carnot')].reset_index()
voisinages["Voisinage rue"] = voisinages["HISTORIQUE"]
voisinages = pd.merge(voisinages.loc[:, voisinages.columns != 'HISTORIQUE'], 
                      df.loc[(~df['HISTORIQUE'].str.startswith("Voisinage")) & (~df['HISTORIQUE'].str.startswith("Voir")), ['HISTORIQUE', 'NOM de VOIE']], 
                      how='left', 
                      on='NOM de VOIE')

voisinages_dup = voisinages[voisinages['Libellé VOIE'].duplicated(keep=False)]
voisinages_dup['NOM de VOIE'].unique()

#On corrige à la main les doublons pour les personnes ; pour le reste, on prendra juste le 1er.
def correction_main(personnes, type_de_voie):
    for personne in personnes:
        df.loc[
            (df['NOM de VOIE']==personne), 'HISTORIQUE'] = df.loc[
            (df['NOM de VOIE']==personne) & (df['TYPE DE VOIE']==type_de_voie), 'HISTORIQUE'].values

correction_main(['anatole france', 'molière', 'carnot'], 'avenue')
correction_main(['voltaire'], 'boulevard')
correction_main(['chartres', 'orléans'], 'galerie')
correction_main(['falguière', 'dauphine', 'sorbonne'], 'rue')
correction_main(['montebello'], 'quai')
# exception car il y a 2 montmorency :
df.loc[df['Libellé VOIE']=='avenue de Montmorency', 'HISTORIQUE'] = df.loc[df['Libellé VOIE']=='boulevard de Montmorency', 'HISTORIQUE'].values

# On a corrigé directement dans le df, donc on supprime ces valeurs de voisinages
voisinages = voisinages[~voisinages['NOM de VOIE'].isin(['anatole france', 'molière', 'carnot', 'voltaire', 'chartres', 'orléans', 'falguière', 'dauphine', 'sorbonne', 'montebello', 'montmorency'])]

# Suppression doublons 
voisinages = voisinages.drop_duplicates(subset='index', keep="first")
voisinages = voisinages.set_index("index")
voisinages.index.name=None

df = df.merge(voisinages.loc[:, ['HISTORIQUE']], how='left', left_index=True, right_index=True)
df['HISTORIQUE']=np.where(df['HISTORIQUE_y'].isna(), df['HISTORIQUE_x'], df['HISTORIQUE_y'])
df = df.drop(columns=['HISTORIQUE_x', 'HISTORIQUE_y'])

df.columns

df.to_csv('Data/voies_paris_labelisé.csv', sep=';')
