==> 4_Graphs.ipynb <==
import pandas as pd
import urllib
import os
import string
import re
import numpy as np

df=pd.read_csv("Data/df_final.csv", sep=";", index_col=0)
df.head()

# 1. Arrondissements

## 1.1. Clean

On cherche à compter les rues féminines et masculines par arrondissement. Or, certaines rues sont à la frontière entre plusieurs arrondissements. On va considérer que ces rues sont alors dans les 2 arrondissements - elles seront alors comptées plusieurs fois au total, mais une fois dans chaque arrondissement.

df_arrondissements = pd.concat([pd.Series(row['Sexe'], row['ARRONDISSEMENT'].split(','))              
                    for _, row in df.iterrows()]).reset_index().rename(columns={'index' : 'Arrondissement', 0:'Sexe'})

df_arrondissements['Arrondissement'] = df_arrondissements['Arrondissement'].apply(lambda x: x.upper())
removed_expressions = [' ', 'E', 'R']
for expression in removed_expressions:
    df_arrondissements['Arrondissement'] = df_arrondissements['Arrondissement'].apply(lambda x: x.replace(expression, ''))
df_arrondissements = df_arrondissements.loc[~ df_arrondissements['Arrondissement'].str.startswith("LIMIT")]

def from_roman(num):
    roman_numerals = {'I':1, 'V':5, 'X':10, 'L':50, 'C':100, 'D':500, 'M':1000}
    try:
        result = 0
        for i,c in enumerate(num):
            if (i+1) == len(num) or roman_numerals[c] >= roman_numerals[num[i+1]]:
                result += roman_numerals[c]
            else:
                result -= roman_numerals[c]
    except:
        result = int(num)
    return result
df_arrondissements['Arrondissement_num'] = df_arrondissements['Arrondissement'].apply(lambda x: from_roman(x))
df_arrondissements=df_arrondissements.loc[df_arrondissements['Arrondissement_num']!=0]

## 1.2. Nombre de rues / sexe par arrondissement

df_arrondissements = pd.crosstab(df_arrondissements['Arrondissement_num'], df_arrondissements['Sexe']).reset_index()
df_arrondissements.columns.name=None

df_arrondissements['Total'] = df_arrondissements[['M', 'F']].sum(axis=1)

df_arrondissements['F_percent'] = df_arrondissements['F']/df_arrondissements['Total']*100

df_arrondissements[['Arrondissement_num', 'F_percent']].to_csv('data/df_arrondissements.csv')

# 2. Quartiers de Paris

## 2.1. Clean

df_quartier = df
# On enlève le point si le quartier finit par un point
pattern_finish_point = re.compile('\.$')
df_quartier['Quartier'] = df_quartier['QUARTIER'].apply(lambda x: pattern_finish_point.sub('', x))

for punctuation in ['.', ',', ';'] :
    df_quartier = pd.concat([pd.Series(row['Sexe'], row['Quartier'].split(punctuation))              
                    for _, row in df_quartier.iterrows()]).reset_index().rename(columns={'index' : 'Quartier', 0:'Sexe'})

# "Sur le territoire de..." à enlever, ce n'est pas un district
df_quartier = df_quartier[~df_quartier['Quartier'].str.contains('Sur le|Situé sur')]

# Enlever espace au début ; mettre en majuscule ; ajouter '-' entre les mots plutôt que espaces
df_quartier['Quartier'] = df_quartier['Quartier'].apply(lambda x: x.strip().title().replace(' ', '-').replace('--', '-'))

# Correction à la main des erreurs qu'il reste
corrections = {'Folies-Méricourt' : 'Folie-Méricourt',
              "Goutte-D\x92Or" : "Goutte-D'Or", 
              "Hôpit" : "Hôpital-Saint-Louis", 
              "La-Villette" : "Villette", 
              "Plaine-De-Monceau" : "Plaine-De-Monceaux", 
              "Saint-Germain-Des-Pré" : "Saint-Germain-Des-Prés", 
              "Épinettes" : "Epinettes", 
              "Les-Halles" : 'Halles',
              "Chapelle" : "La Chapelle", 
              "Saint-Germain-L'Auxerrois" : "St-Germain-L'Auxerrois", 
              'Lambert' : 'Saint-Lambert'}

for erreur in corrections.keys():
    df_quartier.loc[df_quartier['Quartier']==erreur, 'Quartier'] = corrections.get(erreur)

## 2.2. Nombre de rues/sexe par quartier

df_quartier = pd.crosstab(df_quartier['Quartier'], df_quartier['Sexe']).reset_index()
df_quartier.columns.name=None

df_quartier['Total'] = df_quartier[['M', 'F']].sum(axis=1)

df_quartier['F_percent'] = df_quartier['F']/df_quartier['Total']*100

df_quartier[['Quartier', 'F_percent']].to_csv('data/df_quartier.csv')

# 3. Professions

## 3.1. Wordsclouds femmes vs hommes

import stylecloud

for element in ['[', ']', '"', "'"]:
    df.loc[~df['JOBS_list'].isna(), 'JOBS_list'] = df.loc[~df['JOBS_list'].isna(), 'JOBS_list'].apply(lambda x: x.replace(element, ""))
df.loc[~df['JOBS_list'].isna(), 'JOBS_list'] = df.loc[~df['JOBS_list'].isna(), 'JOBS_list'].apply(lambda x: x.split(','))

jobs_list_w = [item for sublist in df.loc[(~df['JOBS_list'].isna()) & (df['Sexe']=='F'), 'JOBS_list'].tolist() for item in sublist]
jobs_list_w = [x.strip(' ') for x in jobs_list_w]

jobs_list_m = [item for sublist in df.loc[(~df['JOBS_list'].isna()) & (df['Sexe']=='M'), 'JOBS_list'].tolist() for item in sublist]
jobs_list_m = [x.strip(' ') for x in jobs_list_m]

text_f=' '.join(jobs_list_w)
text_m=' '.join(jobs_list_m)
file = open("Wordclouds/text_f.txt",'w',  encoding="utf-8") 
file.write(text_f) 
file.close()
file = open("Wordclouds/text_m.txt",'w',  encoding="utf-8") 
file.write(text_m) 
file.close()

stylecloud.gen_stylecloud(file_path="Wordclouds/text_f.txt", 
                        icon_name="fas fa-female", 
                        palette='cartocolors.sequential.Magenta_7')

stylecloud.gen_stylecloud(file_path="Wordclouds/text_m.txt", 
                        icon_name="fas fa-male", 
                        palette='cmocean.sequential.Ice_15_r')

## 3.2. Cleaner les données de jobs?

jobs_list = [item for sublist in df.loc[~df['JOBS_list'].isna(), 'JOBS_list'].tolist() for item in sublist]
jobs_list = set([x.strip(' ') for x in jobs_list])

# Résistants : résistant, otage, résistance, 
# syndicaliste : syndicaliste, militant
# féministe : féministe,
# Titre de nobelesse : comte, marquis, duc, vicomte, seigneur, baron


# homme politique : homme politique, femme politique, maire, conseiller municipal, conseiller, conseillère, préfet, homme d'etat, homme d'état, 
                #homme detat, homme détat, ministre, sénateur, empereur, roi, président, prince, ambassadeur, diplomate, 
                #député, magistrat
# Justice : avocat, jurisconsulte, greffier, juriste, notaire
# Proprétaire : propriétaire, proprietaire, entrepreneur, concessionnaire, gérant
# Artisan/manufacture : manufacturier, industriel, apothicaire, ciseleur, fondeur, graveur, mécanicien, cuisinier, gastronome,
        #horticulteur, imprimeur

# Intellectuel : écrivain, écrivaine, romancier, lithographe,  littérateur, linguiste, anthropologue, érudit, 
            #philosophe, lexicographe, poète, sociologue, sociologiste, économiste, critique, auteur, auteure, dramaturge, 
            #essayiste, homme de lettres, auteur, savant, historien, journaliste, homme de lettres, femme de lettres, calligraphe,
            #chroniqueur, fabuliste, helléniste, instituteur, institutrice, poétesse
            
# Artiste : peintre, graveur, compositeur, compositrice, interprète, dessinateur, dessinatrice, chansonnier, chansonnière, 
            #tragédien, statuaire, sculpteur, danseur, danseuse, décorateur, femme peintre, illustrateur
            # architecte, acteur, actrice, comédien, comédienne, musicien, clarinettiste, guitariste,
            #saxophoniste, violoniste, chef d'orchestre, cantatrice, chanteur, chanteuse, caricaturiste, chorégraphe, cinéaste,
            #compositrice, facteur, humoriste, jardinier
        
# Aviateur : aviateur, aviatrice, pilote, corsaire, marin, navigateur
# Sportif: champion, gymnastique
# Militaire : général, colonel, militaire, amiral, maréchal, officier, capitaine, sergent, lieutenant, gardien de la paix, 
                # homme de guerre
            
# Scientifique : mathématicien, physicien, chimiste, minéarologiste, botaniste, explorateur, ingénieur, inventeur, 
            #aéronaute, naturaliste, voyageur, géologue, agronome, alchimiste, bactériologiste, savant, géographe,
            # archéologue,démographe, ethnographe, ethnologue, géomètre, informaticienne, microbiologiste
# Médecin : docteur, médecin, chirurgien, neurologue, cancérologue, anatomiste, infirmière
# Entourage royal : argentier, chancelier
# Religion : saint, prêtre, pretre, archevêque, archeveque, cardinal, rabbin, jésuite, curé, chef protestant, aumônier,
            #bénédictin, dominicain


#df.loc[(~df['HISTORIQUE'].isna()) & df['HISTORIQUE'].str.contains('intendant'), 'HISTORIQUE'].tolist()

# 4. Dates arrêtés

Y a -t-il une évolution dans le temps du choix des noms donnés aux rues? On regarde les arrêtés depuis 1980.

df.loc[~df['DATE ARRET'].isna(), 'year_arret'] = df.loc[~df['DATE ARRET'].isna(), 'DATE ARRET'].apply(lambda x: int(x[:4]))

df_dates_arret = df.loc[(df['bool_personne']==1) & (df['year_arret']>=1980), ['bool_personne', 'year_arret', 'Sexe']]
pd.crosstab(df_dates_arret['year_arret'], df_dates_arret['Sexe']).to_csv('Data/Date_arret.csv')

# 5. Dates

df.loc[(~df['HISTORIQUE'].isna()) & (df['HISTORIQUE'].str.contains('\(')), 
       'Parenthèses'] = df.loc[(~df['HISTORIQUE'].isna()) & (df['HISTORIQUE'].str.contains('\(')), 
                               'HISTORIQUE'].apply(lambda x: re.search(r'\((.*?)\)',x).group(1))
df.loc[~df['Parenthèses'].isna(), 'Date_fin'] = df.loc[~df['Parenthèses'].isna(), 'Parenthèses'].apply(lambda x: re.search(r'\d{4}(?!\d)', x))

df['Dates_parenthèses'] = df['Parenthèses'].astype(str).apply(lambda x: re.findall('\d+', x))

df.loc[(~df['HISTORIQUE'].isna()), 'Dates_historique'] =  df.loc[(~df['HISTORIQUE'].isna()), 'HISTORIQUE']\
                                                                            .apply(lambda x: re.findall(r'\d+', x))

df.loc[~df['Dates_historique'].isna(), 'Dates_historique'] = df.loc[~df['Dates_historique'].isna(), 'Dates_historique'].apply(lambda x: [i for i in x if len(i)==4])


df.loc[(df['bool_personne']==1) & (~df['Dates_historique'].isna()),'Dates_historique'] = df.loc[(df['bool_personne']==1) & (~df['Dates_historique'].isna()),'Dates_historique'].apply(lambda x: x[:2])

df.loc[(df['bool_personne']==1) & (~df['Dates_historique'].isna()),'nb_dates_historique']=df.loc[(df['bool_personne']==1) & (~df['Dates_historique'].isna()),'Dates_historique'].apply(lambda x: len(x))

df['naissance'] = df.loc[(df['bool_personne']==1) & (df['nb_dates_historique']==2),'Dates_historique'].apply(lambda x: x[0])
